fix products table schema

+++ src/database/migrations/2024_06_07_161843_create_stock_table.php
@@ -15,7 +15,7 @@ class CreateStockTable extends Migration
     {
         Schema::create('stock', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('product_id');
+            $table->mediumInteger('product_id');
             $table->integer('qty');
             $table->timestamps();
             // Foreign key constraint
