+++ src/database/migrations/2024_06_24_063215_create_transaction_details_table.php
@@ -10,7 +10,7 @@ class CreateTransactionDetailsTable extends Migration
     {
         Schema::create('transaction_details', function (Blueprint $table) {
             $table->unsignedBigInteger('transaction_id');
-            $table->unsignedBigInteger('product_id');
+            $table->mediumInteger('product_id');
             $table->unsignedBigInteger('unit');
             $table->integer('qty');
